+++ Backend/src/apps/slot/views.py
@@ -38,5 +38,8 @@ class SlotView(viewsets.GenericViewSet):
 
   
    def deleteSlot(self, request, id):
-       Slot.objects.get(id=id).delete()
-       return JsonResponse({'message': 'Eliminada Correctamente'}, status=status.HTTP_204_NO_CONTENT)
+       slot_data = request.data
+       slot_serializer = SlotSerializer(data=slot_data)
+       if (slot_serializer.is_valid(raise_exception=True)):   
+           Slot.objects.get(id=id).delete()
+       return JsonResponse({'message': 'Slot eliminado Correctamente', "Slot": slot_serializer.data}, status=status.HTTP_204_NO_CONTENT)
